enhance: custom emoji static data 분리
